delete media usage entry

+++ app/src/main/java/com/example/internetapi/models/Media.kt
@@ -1,6 +1,8 @@
 package com.example.internetapi.models
 
+import java.math.BigDecimal
+
 data class MediaType(val id: Int, val name: String)
 data class MediaTypeRequest(val mediaName: String)
-data class MediaUsage(val id: Int, val year: Int, val month: Int, val meterRead: Double)
-data class MediaRegisterRequest(val mediaType: Int, val meterRead: Double, val year: Int, val month: Int)
+data class MediaUsage(val id: Int, val year: Int, val month: Int, val meterRead: BigDecimal)
+data class MediaRegisterRequest(val mediaType: Int, val meterRead: BigDecimal, val year: Int, val month: Int)
